Cast strlen to an int to fix compiler warning.
Since the value that comes in is a long, there is a conflict that
assumes and automatically converts the long to an int and causes the
warning. Since the size of the string most likely won't be that large
(obviously it can be considering the domain, but not in the intended use
of this lib right now) it can be casted to an int explicitly.

+++ ext/xpather/xpather.c
@@ -111,7 +111,7 @@ VALUE constructor(VALUE self, VALUE xmlStr)
   VALUE argv[1];
   VALUE t_data;
 
-  doc = xmlParseMemory(xmlCStr, strlen(xmlCStr));
+  doc = xmlParseMemory(xmlCStr, (int)strlen(xmlCStr));
   if (doc == NULL) {
     fprintf(stderr, "Error: unable to parse xml\n");
     return Qnil;
